app/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
motor
pymongo==4.0.1
bson
uvicorn[standard]==0.15.0
python-dotenv
python-jose[cryptography]
python-multipart

